requirements/staging.txt
-r common.txt
gunicorn==0.14.6
django-storages==1.1.5
boto==2.5.2
django-debug-toolbar==0.9.4
